feat(wintertc): migrate structuredClone from boa_runtime
Move the structuredClone implementation and its supporting JsValueStore
serialization core (the store module) from boa_runtime into boa_wintertc,
following the same move-and-re-export pattern used for atob/btoa (#5418).

- core/runtime/src/clone and core/runtime/src/store are deleted; the code
  now lives only in boa_wintertc.
- boa_runtime re-exports both (pub use boa_wintertc::{clone, store}) so the
  public paths boa_runtime::clone and boa_runtime::store are unchanged, and
  boa_runtime::message keeps compiling against crate::store.
- StructuredCloneExtension delegates to boa_wintertc::clone::register.
- rustc-hash moves with store into boa_wintertc's dependencies; runtime keeps
  it because console still uses it.

store must be public in boa_wintertc so boa_runtime can re-export it for the
message module; net public surface across both crates is unchanged.
