+++ core/runtime/src/extensions.rs
@@ -54,7 +54,7 @@ pub struct StructuredCloneExtension;
 
 impl RuntimeExtension for StructuredCloneExtension {
     fn register(self, realm: Option<Realm>, context: &mut Context) -> JsResult<()> {
-        crate::clone::register(realm, context)
+        boa_wintertc::clone::register(realm, context)
     }
 }
