+++ core/wintertc/Cargo.toml
@@ -13,6 +13,7 @@ rust-version.workspace = true
 [dependencies]
 boa_engine.workspace = true
 base64.workspace = true
+rustc-hash = { workspace = true, features = ["std"] }
 
 [dev-dependencies]
 indoc.workspace = true
